Fix Devel::PeekPoke on cygwin perls
Apparently how data is written into a union larger than the variable type is
entirely compiler/paltform dependent. In cases when ivsize > ptrsize we need
to find out exactly which part of the (ivsize sized) svu_pv will hold the
pointer before we go overwriting it. The previous assumption this offset is
always 0 turned out to be false.

Add some extra test suite diags while we are at it as well.